+++ dashboard/package.json
@@ -19,6 +19,9 @@
 		"@smui/card": "^7.0.0-beta.14",
 		"@smui/data-table": "^7.0.0-beta.14",
 		"@smui/form-field": "^7.0.0-beta.14",
+		"@smui/paper": "^7.0.0-beta.14",
+		"@smui/tab": "^7.0.0-beta.14",
+		"@smui/tab-bar": "^7.0.0-beta.14",
 		"@smui/textfield": "^7.0.0-beta.14",
 		"@sveltejs/adapter-auto": "^2.0.0",
 		"@sveltejs/kit": "^1.5.0",
