+++ dashboard/src/routes/+page.svelte
@@ -1,52 +1,12 @@
 <script lang="ts">
 	import '../app.css';
-	import Button, { Label } from '@smui/button';
-	import Textfield from '@smui/textfield';
-	import DataTable, { Head, Body, Row, Cell } from '@smui/data-table';
+	import CowInputField from '$lib/components/cowInputField.svelte';
+	import Tab, { Label } from '@smui/tab';
+  	import TabBar from '@smui/tab-bar';
+	import CowTable from '$lib/components/cowTable.svelte';
 
-	const formFields = ["Cow's  Name", "Farm Name", "Cow's Genetic", "Father's Name", "Father's Genetic", "Mother's Name", "Mother's Genetic"];
-
-	let tempCow = ["", "", "", "", "", "", ""];
-	let birthdate:number = new Date().getTime();
-	let weightAtBirth = 0;
-
-	let cowsData: any;
-	const showCows = async () => {
-		cowsData = await fetch('/api/getCows').then(async res => res.json());
-		console.log(cowsData);
-	}
-
-	const createCow = () => {
-		const data = {
-			name: tempCow[0],
-			farmID: tempCow[1],
-			genetic: tempCow[2],
-			birthDate: birthdate,
-			weightAtBirth: weightAtBirth,
-			fatherName: tempCow[3],
-			fatherGenetic: tempCow[4],
-			motherName: tempCow[5],
-			motherGenetic: tempCow[6]
-		}
-		// console.log(JSON.stringify(data));
-		postData('/api/cow', data);
-	}
-
-	const postData = (url: string, data: any) => {
-		return fetch(url, {
-			method: 'POST',
-			body: JSON.stringify(data),
-			headers: {
-				'Content-Type': 'application/json'
-			}
-		})
-		.then(async (res) => {
-			const result = await res.json()
-			console.log(`Result : `,result)
-			return result
-		})
-		.catch(err => console.log(err));
-	}
+	let mode = "Insert";
+	let table = 'Cow';
 </script>
 
 <svelte:head>
@@ -54,59 +14,26 @@
   <meta name="description" content="Dashboard for Pdairy's PostgreSQL" />
 </svelte:head>
 
-<h1 class="text-4xl font-bold m-auto">Dashboard</h1>
-<!-- <p class="text-xl m-auto">Welcome to the dashboard!</p> -->
-
-<div class="flex flex-row">
-	<div class="flex flex-col p-3 border-zinc-800 border-2 mx-2 rounded-md">
-		{#each formFields as field, index}
-			<Textfield label={field} bind:value={tempCow[index]} class="mt-2" input$maxlength={30} />	
-		{/each}
-		<Textfield label="Birth Date" type="datetime-local" bind:value={birthdate} class="mt-2"/>
-		<Textfield label="Weight At Birth" type="number" bind:value={weightAtBirth} class="mt-2"/>
-		
-		<Button on:click={createCow} class="mt-5" variant="raised">
-			<Label>Insert Cow</Label>
-		</Button>
-	</div>
-</div>
-<div class="w-full m-2">
-	<Button on:click={showCows} class="mt-5" variant="raised">
-		<Label>Show Cows</Label>
-	</Button>
-	{#if cowsData?.length > 0}
-	<DataTable stickyHeader table$aria-label="User list" style="width: 100%;">
-		<Head>
-		  <Row>
-			<Cell numeric>ID</Cell>
-			<Cell>Name</Cell>
-			<Cell>Genetic</Cell>
-			<Cell>Birth Date</Cell>
-			<Cell>Weight At Birth</Cell>
-			<Cell>Father's Name</Cell>
-			<Cell>Father's Genetic</Cell>
-			<Cell>Mother's Name</Cell>
-			<Cell>Mother's Genetic</Cell>
-		  </Row>
-		</Head>
-		<Body>
-		  {#each cowsData as data, index}
-			<Row>
-			  	<Cell numeric>{data.ID}</Cell>
-				<!-- <Cell>{data.id}</Cell> -->
-				<Cell>{data.name}</Cell>
-				<Cell>{data.genetic}</Cell>
-				<Cell>{data.birthDate}</Cell>
-				<Cell>{data.weightAtBirth}</Cell>
-				<Cell>{data.fatherName}</Cell>
-				<Cell>{data.fatherGenetic}</Cell>
-				<Cell>{data.motherName}</Cell>
-				<Cell>{data.motherGenetic}</Cell>
-			</Row>
-		  {/each}
-		</Body>
-	</DataTable>
-	{:else}
-		<p>No cows to show</p>
+<div class="flex flex-col items-center max-w-screen-md m-auto">
+	<h1 class="text-4xl font-bold m-auto">Dashboard</h1>
+	<!-- <p class="text-xl m-auto">Welcome to the dashboard!</p> -->
+	<TabBar tabs={['Cow', 'Farm', 'Milk', "Breeding"]} let:tab bind:active={table} class="w-fit self-start">
+		<Tab {tab} minWidth>
+		<Label>{tab}</Label>
+		</Tab>
+	</TabBar>
+	<TabBar tabs={['Insert', 'View']} let:tab bind:active={mode} class="w-fit self-start">
+		<Tab {tab} minWidth>
+		<Label>{tab}</Label>
+		</Tab>
+	</TabBar>
+	{#if mode === 'Insert'}
+		{#if table === 'Cow'}
+			<CowInputField />
+		{/if}
+	{:else if mode === 'View'}
+		{#if table === 'Cow'}
+			<CowTable />
+		{/if}
 	{/if}
 </div>
