+++ dashboard/tsconfig.json
@@ -8,7 +8,8 @@
 		"resolveJsonModule": true,
 		"skipLibCheck": true,
 		"sourceMap": true,
-		"strict": true
+		"strict": true,
+		"ignoreDeprecations" : "5.0",
 	}
 	// Path aliases are handled by https://kit.svelte.dev/docs/configuration#alias
 	//
