--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "npm": "^9.8.0"
-  }
-}
